Implemented EXTREMELY janky method of reading multiple messages
I need to go to bed so I've patched it over with a terrible but working solution. I need to figure out how to actually get the number of bytes that are left in the buffer after reading some out. It always seems to tell me how many bytes were in there at the beginning. So for now I will just manually subtract the number of bytes read from a counter.